pyproject.toml
[project]
name = "kujiya"
version = "1.0.0"
description = "the future of ai in africa"
dependencies = [
    "flask",
"python-dotenv"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
